Error in varing layout, in the vertex shaders the outColor was deleted, but was need in the fragment shaders

+++ a.sh
@@ -1,12 +1,14 @@
 #!/bin/bash
 vert=$(glslang vShaders.vert)
 frag=$(glslang fShaders.frag)
+geom=$(glslang gShaders.geom)
 
-if [[ "$vert" = "" && "$frag" = ""  ]]; then
+if [[ "$vert" = "" && "$frag" = ""  && "$geom" = ""  ]]; then
   rm build/*
   make
   ./build/final
 else 
   echo $vert
   echo $frag
+  echo $geom
 fi
